separate components: FriendLists - lists all friends Friend - single friend NewFriend - form to add a friend
